+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 1.1.0
 
 - Added LOD Bias setting
+- Added Upscaling Filter setting
 - Improved hold item in one hand
+- Improved hands syncronization
 - Updated README
 
 ## 1.0.0